feat: add projects and messages contexts
